Common parent class for card and wallet check enrollment
